UnbindableDir is a convenience class for creating a temp dir and its Unbinder
